+++ app/static/css/style.css
@@ -1,94 +1,130 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@500&display=swap');
 
-*{
+* {
     box-sizing: border-box;
     margin: 0;
     padding: 0;
 }
 
+button {
+    border: none;
+    outline: none;
+}
+
 body {
     margin: 0;
-    height: 100vh; /* 100% of viewport height */
-    display: flex;
-    background-color: #000000;
+    height: 100vh;
+    /* 100% of viewport height */
+    background-color: black;
 }
 
 .Main-container {
-    font-family: 'Roboto', sans-serif;
-    font-weight: 500;
     width: 100%;
     height: 100%;
+    font-family: 'Roboto', sans-serif;
+    font-weight: 500;
     display: flex;
-    align-items: center ;
+    align-items: center;
+    justify-content: center;
     flex-direction: column;
 }
 
-.img-title {
-    width: 30%;
-    height: auto;
-    max-width: 100%;
-    max-height: 100%;
-    pointer-events: none;
-}
-
-@media (max-width: 768px) {
-    .img-title {
-        width: 100%;
-    }
-}
-
-.input {
+.main-page1 {
     width: 80%;
-    height: 20%;
-    flex: 1;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-}
-
-.search_bar {
-    margin-top: 40px;
+    height: 10vh;
     display: flex;
+    /* Enable flexbox for horizontal arrangement */
+    justify-content: center;
+    /* Center content horizontally */
     align-items: center;
-    border-bottom: .1rem inset rgba(255, 255, 255, 0.5);
+    margin-bottom: 15%;
 }
 
-.search_bar input {
+.main-page1 input[type="text"],
+.main-page1 .reset,
+.main-page1 .proceed,
+.main-page1 .paste {
     font-family: 'Roboto', sans-serif;
     font-size: 1.5rem;
-    background: none;
-    flex: 1;
-    border: 0;
-    outline: none;
-    color: white;
-    padding: 12px 8px;
+    height: 40px;
+    /* Set same height for all elements */
+    margin-right: 10px;
+    /* Add spacing between elements */
+    margin-top: 10px;
 }
 
-#reset {
-    background-color: black;
-    border: 0;
-    cursor: pointer;
+.main-page1 input[type="text"] {
+    width: 40%;
+    color: white;
+    background: none;
+    border: none;
+    border-bottom: .1rem inset rgba(255, 255, 255, 0.5);
+    border-right: .1rem inset rgba(255, 255, 255, 0.5);
+    outline: none;
+    font-size: 18px;
+    line-height: 20px;
     padding: 5px 8px;
 }
 
-#reset img {
-    background-color: white;
-    height: 50%;
-    width: 70%;
+.main-page1 .reset img,
+.main-page1 .proceed img,
+.main-page1 .paste img {
+    background-color: black;
+    height: 40px;
+    width: auto;
+    border: 0;
+    cursor: pointer;
+    vertical-align: middle;
+    /* Align images vertically within buttons */
 }
 
-.extract {
-    background-color: black;
-    border: none;
-    display: inline-block;
+.option {
+    width: 80%;
+    height: 10vh;
+    display: flex;
+    /* Enable flexbox for horizontal arrangement */
+    justify-content: center;
+    /* Center content horizontally */
+    align-items: center;
+}
+
+.option button {
+    font-family: 'Roboto', sans-serif;
+    font-size: 18px;
+    line-height: 20px;
+    height: 50px;
+    background-color: rgb(44, 44, 44);
+    color: white;
+    border: 1px solid rgba(84, 82, 82, 0.5);
+    padding: 0 20px;
+    /* Set same height for all elements */
     cursor: pointer;
 }
 
-.extract img {
-    height: 20%;
-    width: 20%;
+.option #advanced-button {
+    border-left: none;
+    border-right: none;
 }
 
+.option #video-button {
+    border-right: none;
+    border-radius: 10px 0px 0px 10px;
+}
+
+.option #audio-button {
+    border-left: none;
+    border-radius: 0px 10px 10px 0px;
+}
+
+.option button.selected {
+    color: black;
+    background-color: #f5f5f5;
+    border-color: #000000;
+    box-shadow: 0 0 2px rgba(0, 0, 0, 0.15);
+}
+  
+
+/*
 .video-info {
     font-family: 'roboto', sans-serif;
     font-weight: 500;
@@ -121,4 +157,4 @@ body {
     width: 100%;
     height: 100%;
     object-fit: contain;
-}
+} */
