+++ app/templates/index.html
@@ -6,7 +6,6 @@
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <link rel="icon" href="../static/images/favicon.ico" type="image/x-icon" sizes="32x32">
         <link rel="stylesheet" type="text/css" href="../static/css/style.css" />
-        <script src="../static/js/script.js"></script>
     </head>
     <body>
         <div class = "Main-container">
